+++ tests/store/tracking/test_sqlalchemy_store.py
@@ -39,7 +39,6 @@ from mlflow.protos.databricks_pb2 import (
     TEMPORARILY_UNAVAILABLE,
     ErrorCode,
 )
-from mlflow.store.artifact.artifact_repository_registry import get_artifact_repository
 from mlflow.store.db.db_types import MSSQL, MYSQL, POSTGRES, SQLITE
 from mlflow.store.db.utils import (
     _get_latest_schema_revision,
@@ -2460,60 +2459,6 @@ class TestSqlAlchemyStore(unittest.TestCase, AbstractStoreTest):
 
         results = self.store._search_datasets([exp_id])
         assert len(results) == 1000
-
-    @pytest.mark.skipif(
-        "MLFLOW_SKINNY" in os.environ,
-        reason="Skinny does not support the np or pandas dependencies",
-    )
-    def test_create_promptlab_run(self):
-        exp_id = self.store.create_experiment("test_create_promptlab_run")
-        run = self.store._create_promptlab_run(
-            experiment_id=exp_id,
-            run_name="my_promptlab_run",
-            tags=[],
-            prompt_template="my_prompt_template",
-            prompt_parameters=[Param("prompt_param_key", "prompt_param_value")],
-            model_route="my_route",
-            model_parameters=[Param("temperature", "0.1")],
-            model_input="",
-            model_output_parameters=[Param("output_param_key", "output_param_value")],
-            model_output="my_output",
-            mlflow_version="1.0.0",
-            user_id="user",
-            start_time=1,
-        )
-        assert run.info.run_id is not None
-        assert run.info.status == RunStatus.to_string(RunStatus.FINISHED)
-
-        assert run.data.params["prompt_template"] == "my_prompt_template"
-        assert run.data.params["model_route"] == "my_route"
-        assert run.data.params["temperature"] == "0.1"
-
-        assert run.data.tags["mlflow.runName"] == "my_promptlab_run"
-        assert (
-            run.data.tags["mlflow.loggedArtifacts"]
-            == '[{"path": "eval_results_table.json", "type": "table"}]'
-        )
-        assert run.data.tags["mlflow.runSourceType"] == "PROMPT_ENGINEERING"
-        assert run.data.tags["mlflow.log-model.history"] is not None
-
-        # list the files in the model folder
-        artifact_location = run.info.artifact_uri
-        artifact_repo = get_artifact_repository(artifact_location)
-
-        artifact_files = [f.path for f in artifact_repo.list_artifacts()]
-        assert "eval_results_table.json" in artifact_files
-        assert "model" in artifact_files
-
-        model_files = [f.path for f in artifact_repo.list_artifacts("model")]
-        assert "model/MLmodel" in model_files
-        assert "model/python_env.yaml" in model_files
-        assert "model/conda.yaml" in model_files
-        assert "model/requirements.txt" in model_files
-        assert "model/input_example.json" in model_files
-
-        # try to load the model
-        mlflow.pyfunc.load_model(os.path.join(artifact_location, "model"))
 
     def test_log_batch(self):
         experiment_id = self._experiment_factory("log_batch")
